+++ test/libponyrt/lang/error.cc
@@ -3,6 +3,8 @@
 
 #include <pony.h>
 
+#include <exception>
+
 TEST(ErrorTest, PonyTry)
 {
   auto cb_success = [](void*){};
@@ -11,3 +13,16 @@ TEST(ErrorTest, PonyTry)
   ASSERT_TRUE(pony_try(cb_success, nullptr));
   ASSERT_FALSE(pony_try(cb_error, nullptr));
 }
+
+// Doesn't work on windows. The C++ code doesn't catch C++ exceptions if they've
+// traversed a pony_try() call. This is suspected to be related to how SEH and
+// LLVM exception code generation interact.
+// See https://github.com/ponylang/ponyc/issues/2455 for more details.
+#ifndef PLATFORM_IS_WINDOWS
+TEST(ErrorTest, PonyTryCantCatchCppExcept)
+{
+  auto cb_cppthrow = [](void*){ throw std::exception{}; };
+
+  ASSERT_THROW((pony_try(cb_cppthrow, nullptr)), std::exception);
+}
+#endif